fix multi-monitor bugs on using menu/distribute/fill-space
Previously considered windows from the current workspace.
So when there was 2 windows in #1 monitor and 4 windows in the #2 monitor, it arranged a 6 places grid for the #1 monitor. When a distribute command was used it arranged following this grid but the windows remained in respective monitor, 2 in #1 and 4 in #2.

Now consider windows from the monitor of the current window only.
So the windows from the other monitor in the current workarea are not affected and the grid are calculated correctly.
In the example given, it will calculated a 2 places grid for the #1 monitor.
Teste dwith 2 1920x1280 monitors in ubuntu disco.